+++ NEWS.md
@@ -1,6 +1,7 @@
 # Changes in Version 1.1-2 (BS)
 - DESCRIPTION modified according to CRAN guidelines
 - seed is now an input for GL functions
+- error fixed in the test file of span_bj
 
 # Changes in Version 1.1-1 (BS)
 - Function span_km and span_bj were testing the same